Fixed connection pooling with cancelable timeout (#763)
